# tests/Api.Tests/Endpoints/OpenApi/OpenApiTests.OpenApi_VerifyAsExpected.verified.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Waste Obligations REST API",
    "description": "Manage obligation data in relation to EPR",
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "https://localhost"
    }
  ],
  "paths": {
    "/organisations/{organisationId}/obligations": {
      "get": {
        "tags": [
          "Obligations"
        ],
        "summary": "Obligations for an organisation by year",
        "description": "Returns the obligations for an organisation by organisation ID for the specified year",
        "operationId": "ReadOrganisationObligations",
        "parameters": [
          {
            "name": "organisationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "obligationYear",
            "in": "query",
            "schema": {
              "maximum": 2050,
              "minimum": 2023,
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganisationObligations"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/organisations/{organisationId}/compliance-declarations": {
      "post": {
        "tags": [
          "ComplianceDeclarations"
        ],
        "summary": "Create a compliance declaration",
        "description": "Create a compliance declaration for the specified organisation ID",
        "operationId": "CreateOrganisationComplianceDeclaration",
        "parameters": [
          {
            "name": "organisationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateComplianceDeclarationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplianceDeclaration"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "ComplianceDeclarations"
        ],
        "summary": "Compliance declarations for an organisation by year",
        "description": "Returns the compliance declarations for an organisation by organisation ID for the specified year",
        "operationId": "ReadOrganisationComplianceDeclarations",
        "parameters": [
          {
            "name": "organisationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "obligationYear",
            "in": "query",
            "schema": {
              "maximum": 2050,
              "minimum": 2023,
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganisationComplianceDeclarations"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/organisations/{organisationId}/compliance-declarations/{complianceDeclarationId}": {
      "get": {
        "tags": [
          "ComplianceDeclarations"
        ],
        "summary": "Compliance declaration by ID",
        "description": "Return a compliance declaration by compliance declaration ID",
        "operationId": "ReadOrganisationComplianceDeclaration",
        "parameters": [
          {
            "name": "organisationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "complianceDeclarationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplianceDeclaration"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "ComplianceDeclarations"
        ],
        "summary": "Update compliance declaration by ID",
        "description": "Update a compliance declaration by compliance declaration ID",
        "operationId": "UpdateOrganisationComplianceDeclaration",
        "parameters": [
          {
            "name": "organisationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "complianceDeclarationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateComplianceDeclarationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplianceDeclaration"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Address": {
        "type": "object",
        "properties": {
          "addressLine1": {
            "type": [
              "null",
              "string"
            ]
          },
          "addressLine2": {
            "type": [
              "null",
              "string"
            ]
          },
          "town": {
            "type": [
              "null",
              "string"
            ]
          },
          "county": {
            "type": [
              "null",
              "string"
            ]
          },
          "postcode": {
            "type": [
              "null",
              "string"
            ]
          },
          "country": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "AuditEntry": {
        "type": "object",
        "anyOf": [
          {
            "$ref": "#/components/schemas/AuditEntryReasonAuditEntry"
          },
          {
            "$ref": "#/components/schemas/AuditEntryBase"
          }
        ]
      },
      "AuditEntryBase": {
        "required": [
          "user",
          "timestamp",
          "action"
        ],
        "properties": {
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "action": {
            "type": "string"
          }
        }
      },
      "AuditEntryReasonAuditEntry": {
        "required": [
          "reason",
          "user",
          "timestamp",
          "action"
        ],
        "properties": {
          "reason": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "action": {
            "type": "string"
          }
        }
      },
      "ComplianceDeclaration": {
        "required": [
          "id",
          "organisation",
          "obligationYear",
          "obligationStatus",
          "declarationText",
          "submitterName"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "created": {
            "type": "string",
            "description": "ISO 8601 extended format with offset",
            "format": "date-time"
          },
          "updated": {
            "type": "string",
            "description": "ISO 8601 extended format with offset",
            "format": "date-time"
          },
          "status": {
            "$ref": "#/components/schemas/ComplianceDeclarationStatus"
          },
          "organisation": {
            "$ref": "#/components/schemas/OrganisationRequest"
          },
          "obligationYear": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Obligation year of compliance declaration",
            "format": "int32"
          },
          "obligations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Obligation"
            }
          },
          "obligationStatus": {
            "enum": [
              "Met",
              "NotMet"
            ],
            "type": "string"
          },
          "declarationText": {
            "$ref": "#/components/schemas/LocalizedText"
          },
          "submitterName": {
            "type": "string"
          },
          "audit": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuditEntry"
            }
          }
        }
      },
      "ComplianceDeclarationStatus": {
        "enum": [
          "Submitted",
          "Accepted",
          "Cancelled"
        ]
      },
      "CreateComplianceDeclarationRequest": {
        "required": [
          "organisation",
          "obligationYear",
          "obligationStatus",
          "declarationText",
          "submitterName",
          "user"
        ],
        "type": "object",
        "properties": {
          "organisation": {
            "$ref": "#/components/schemas/OrganisationRequest"
          },
          "obligationYear": {
            "maximum": 2050,
            "minimum": 2023,
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "obligations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Obligation"
            }
          },
          "obligationStatus": {
            "enum": [
              "Met",
              "NotMet"
            ],
            "type": "string"
          },
          "declarationText": {
            "$ref": "#/components/schemas/LocalizedText"
          },
          "submitterName": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "LocalizedText": {
        "required": [
          "text",
          "language"
        ],
        "type": "object",
        "properties": {
          "text": {
            "type": "string"
          },
          "language": {
            "type": "string",
            "description": "Language string for ISO 639-1 (example en) or BCP 47 (example en-GB)"
          }
        }
      },
      "Obligation": {
        "required": [
          "material",
          "tonnages",
          "status"
        ],
        "type": "object",
        "properties": {
          "material": {
            "enum": [
              "Plastic",
              "Glass",
              "Aluminium",
              "Steel",
              "Wood",
              "GlassRemelt",
              "Paper"
            ],
            "type": "string",
            "description": "Name of material"
          },
          "recyclingTarget": {
            "maximum": 1,
            "minimum": 0,
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Recycling target expressed in decimal form",
            "format": "double"
          },
          "tonnages": {
            "$ref": "#/components/schemas/ObligationTonnages"
          },
          "status": {
            "enum": [
              "NoDataYet",
              "Met",
              "NotMet"
            ],
            "type": "string",
            "description": "Status of obligation"
          }
        }
      },
      "ObligationTonnages": {
        "type": "object",
        "properties": {
          "material": {
            "maximum": 2147483647,
            "minimum": 0,
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "awaitingAcceptance": {
            "maximum": 2147483647,
            "minimum": 0,
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "accepted": {
            "maximum": 2147483647,
            "minimum": 0,
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "outstanding": {
            "maximum": 2147483647,
            "minimum": 0,
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "obligated": {
            "maximum": 2147483647,
            "minimum": 0,
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "OrganisationComplianceDeclarations": {
        "type": "object",
        "properties": {
          "complianceDeclarations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ComplianceDeclaration"
            }
          }
        }
      },
      "OrganisationObligations": {
        "type": "object",
        "properties": {
          "obligations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Obligation"
            }
          }
        }
      },
      "OrganisationRequest": {
        "required": [
          "id",
          "name",
          "regulator",
          "regulatorEmail"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "complianceSchemeName": {
            "type": [
              "null",
              "string"
            ]
          },
          "schemeOperatorName": {
            "type": [
              "null",
              "string"
            ]
          },
          "referenceNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "address": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Address"
              }
            ]
          },
          "regulator": {
            "type": "string"
          },
          "regulatorEmail": {
            "type": "string"
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "UpdateComplianceDeclarationRequest": {
        "required": [
          "user"
        ],
        "type": "object",
        "properties": {
          "status": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ComplianceDeclarationStatus"
              }
            ]
          },
          "reason": {
            "type": [
              "null",
              "string"
            ]
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "User": {
        "required": [
          "id",
          "email"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Obligations"
    },
    {
      "name": "ComplianceDeclarations"
    }
  ]
}